backend/requirements.txt
# Django
Django==6.0.2
djangorestframework==3.16.0
django-import-export==4.3.6
django-qsessions==2.1.0
python-decouple==3.8
redis==7.1.0
django-redis==6.0.0
rq==2.6.1
django-rq==3.2.2
django-mongodb-backend==6.0.2

# Deployment
gunicorn==23.0.0
uvicorn[standard]==0.34.0
whitenoise==6.11.0

# Ingestion
requests
feedparser>=6.0
telethon>=1.36
transformers>=5.3
vaderSentiment>=3.3
openai>=2.24.0

# PyTorch CPU-only (avoids pulling CUDA binaries)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu

# Geocoding
geonamescache>=1.3

# Stream data fetching (prices, forex, NOTAM, earthquakes)
yfinance>=0.2.54
